requirements.txt
# Core dependencies
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0

# Deep Learning
tensorflow>=2.8.0
torch>=1.11.0
torchvision>=0.12.0

# Data collection and APIs
requests>=2.28.0
yfinance>=0.1.70
websocket-client>=1.3.0

# Database
sqlite3  # Built into Python
influxdb-client>=1.29.0

# Visualization
streamlit>=1.10.0
plotly>=5.8.0
matplotlib>=3.5.0
seaborn>=0.11.0

# Configuration and utilities
pyyaml>=6.0
python-dotenv>=0.20.0
tqdm>=4.64.0
loguru>=0.6.0

# Testing
pytest>=7.1.0
pytest-cov>=3.0.0

# Development
black>=22.3.0
flake8>=4.0.0
mypy>=0.950

# API and web
fastapi>=0.78.0
uvicorn>=0.17.0

# Time series specific
ta>=0.10.0
statsmodels>=0.13.0

# News and sentiment analysis
textblob>=0.17.1
vaderSentiment>=3.3.2
